refactor: split oversized modules to meet 300-line file-length gate
Split 5 files that exceeded the 300-line limit into focused modules:

- daemon.py (381→149): extract wire protocol + job handling → daemon_protocol.py
- nats/tts_adapter.py (360→255): extract WAV helpers → nats/tts_wav_utils.py
- overlay.py (441→247): extract drawing → overlay_draw.py, audio → overlay_audio.py
- translate.py (342→173): extract ENGINE_CAPS data → engine_caps.py
- markdown.py (416→149): extract directive parsing → _markdown_directives.py

Also:
- Remove 5 now-resolved file exemptions (overlay, markdown, translate, daemon, tts_adapter)
- Remove 2 stale exemptions: nats/sst_adapter.py (<300 lines), src/voicecli/nats (<12 files)
- Add daemon_protocol.py exemption (318 lines, single remaining overage)
- Fix overlay_draw.py Cairo FontSlant/FontWeight enum types
- Add type: ignore for circular-import deferred imports in markdown.py
- Update pyrightconfig.json: ignore overlay + nats paths (no GTK/roxabi_nats stubs)
