+++ src/voicecli/markdown.py
@@ -124,293 +124,26 @@ def strip_markdown(text: str) -> str:
     return text.strip()
 
 
-_COMMENT_RE = re.compile(r"<!--(.+?)-->", re.DOTALL)
+def parse_md_file(path: Path) -> TTSDocument:
+    """Parse a .md file into a TTSDocument."""
+    # Deferred import avoids circular dependency:
+    # _markdown_directives imports from this module at its module level,
+    # which is safe here because markdown.py is fully initialised before
+    # parse_md_file() is ever called.
+    from voicecli._markdown_directives import parse_md_file as _parse  # type: ignore[import-not-found]
 
-# Fields that can appear as <!-- key: value --> inline directives
-_STR_DIRECTIVES = {"instruct", "accent", "personality", "speed", "emotion", "language", "voice"}
-_FLOAT_DIRECTIVES = {
-    "exaggeration",
-    "cfg_weight",
-    "cfg_alpha",
-    "temperature",
-    "top_p",
-    "min_p",
-    "repetition_penalty",
-}
-_INT_DIRECTIVES = {"segment_gap", "crossfade", "flow_steps"}
-_ALL_DIRECTIVES = _STR_DIRECTIVES | _FLOAT_DIRECTIVES | _INT_DIRECTIVES
+    return _parse(path)
 
 
 def _parse_comment_kvs(content: str) -> dict[str, str]:
-    """Parse comma-separated key: value pairs from an HTML comment body.
+    """Re-export from _markdown_directives (used by tests and any external callers)."""
+    from voicecli._markdown_directives import _parse_comment_kvs as _fn  # type: ignore[import-not-found]
 
-    Handles quoted values (with commas inside), unquoted values, single/double quotes.
-    Example: 'emotion: "Passionnée, mais contenue", speed: "Rapide"' →
-             {"emotion": "Passionnée, mais contenue", "speed": "Rapide"}
-    """
-    result: dict[str, str] = {}
-    content = content.strip()
-    i = 0
-    n = len(content)
-
-    while i < n:
-        # Skip whitespace and commas between pairs
-        while i < n and content[i] in (" ", "\t", "\n", "\r", ","):
-            i += 1
-        if i >= n:
-            break
-
-        # Parse key (word chars until colon)
-        key_start = i
-        while i < n and content[i] not in (":", " ", "\t", "\n"):
-            i += 1
-        key = content[key_start:i].strip()
-        if not key:
-            break
-
-        # Skip to colon
-        while i < n and content[i] in (" ", "\t"):
-            i += 1
-        if i >= n or content[i] != ":":
-            break
-        i += 1  # skip colon
-
-        # Skip whitespace after colon
-        while i < n and content[i] in (" ", "\t"):
-            i += 1
-        if i >= n:
-            break
-
-        # Parse value — quoted or unquoted
-        if content[i] in ('"', "'"):
-            quote = content[i]
-            i += 1  # skip opening quote
-            val_start = i
-            while i < n and content[i] != quote:
-                i += 1
-            value = content[val_start:i]
-            if i < n:
-                i += 1  # skip closing quote
-        else:
-            # Unquoted: read until comma or end
-            val_start = i
-            while i < n and content[i] != ",":
-                i += 1
-            value = content[val_start:i].strip()
-
-        result[key] = value
-
-    return result
-
-
-def _parse_directive_value(key: str, raw: str) -> object:
-    """Parse a directive value to its expected type."""
-    raw = raw.strip()
-    # Strip surrounding quotes for string values
-    if len(raw) >= 2 and raw[0] == raw[-1] and raw[0] in ('"', "'"):
-        raw = raw[1:-1]
-    if key in _FLOAT_DIRECTIVES:
-        return float(raw)
-    if key in _INT_DIRECTIVES:
-        return int(raw)
-    return raw
-
-
-def _compose_segment_instruct(seg: Segment, pending_overrides: dict) -> None:
-    """Compose instruct from structured parts if appropriate.
-
-    Rules:
-    - If instruct was explicitly set via inline directive → keep as-is (bypass)
-    - If any structured part was overridden at this section level → compose from all parts
-    - Otherwise, compose from inherited parts if no inherited instruct
-    """
-    parts_overridden = any(k in pending_overrides for k in _INSTRUCT_PARTS)
-    if parts_overridden:
-        composed = compose_instruct(seg.accent, seg.personality, seg.speed, seg.emotion)
-        if composed:
-            seg.instruct = composed
-    elif "instruct" not in pending_overrides and seg.instruct is None:
-        composed = compose_instruct(seg.accent, seg.personality, seg.speed, seg.emotion)
-        if composed:
-            seg.instruct = composed
+    return _fn(content)
 
 
 def _parse_segments(body: str, defaults: dict) -> list[Segment]:
-    """Split body on <!-- key: value --> directives into per-section segments.
+    """Re-export from _markdown_directives (used by tests and any external callers)."""
+    from voicecli._markdown_directives import _parse_segments as _fn  # type: ignore[import-not-found]
 
-    Each section inherits from frontmatter defaults, overridden by inline directives.
-    Consecutive directives accumulate and apply to the text that follows.
-    """
-    # Find all HTML comments and their positions
-    matches = list(_COMMENT_RE.finditer(body))
-
-    # No directives found → single segment with defaults
-    if not matches:
-        text = strip_markdown(body)
-        if text:
-            seg = Segment(text=text, **{k: v for k, v in defaults.items() if k != "text"})
-            _compose_segment_instruct(seg, {})
-            return [seg]
-        return []
-
-    segments: list[Segment] = []
-    pending_overrides: dict = {}
-    prev_end = 0
-
-    for match in matches:
-        # Text between previous position and this comment
-        text_before = body[prev_end : match.start()]
-        stripped = strip_markdown(text_before)
-        if stripped:
-            seg_kwargs = {**defaults, **pending_overrides, "text": stripped}
-            seg = Segment(**seg_kwargs)
-            _compose_segment_instruct(seg, pending_overrides)
-            segments.append(seg)
-            pending_overrides = {}
-
-        # Parse all key-value pairs from this comment
-        kvs = _parse_comment_kvs(match.group(1))
-        for key, raw_value in kvs.items():
-            if key in _ALL_DIRECTIVES:
-                try:
-                    pending_overrides[key] = _parse_directive_value(key, raw_value)
-                except (ValueError, TypeError):
-                    pass
-
-        prev_end = match.end()
-
-    # Remaining text after last directive
-    remaining = strip_markdown(body[prev_end:])
-    if remaining:
-        seg_kwargs = {**defaults, **pending_overrides, "text": remaining}
-        seg = Segment(**seg_kwargs)
-        _compose_segment_instruct(seg, pending_overrides)
-        segments.append(seg)
-
-    return segments
-
-
-def _parse_optional_float(metadata: dict, key: str) -> float | None:
-    if key in metadata:
-        try:
-            return float(metadata[key])
-        except ValueError:
-            pass
-    return None
-
-
-def _parse_optional_int(metadata: dict, key: str) -> int | None:
-    if key in metadata:
-        try:
-            return int(metadata[key])
-        except ValueError:
-            pass
-    return None
-
-
-def parse_md_file(path: Path) -> TTSDocument:
-    """Parse a .md file into a TTSDocument."""
-    content = path.read_text(encoding="utf-8")
-    metadata, body = parse_frontmatter(content)
-
-    known_keys = {
-        "language",
-        "voice",
-        "engine",
-        "instruct",
-        "accent",
-        "personality",
-        "speed",
-        "emotion",
-        "exaggeration",
-        "cfg_weight",
-        "flow_steps",
-        "cfg_alpha",
-        "temperature",
-        "top_p",
-        "min_p",
-        "repetition_penalty",
-        "segment_gap",
-        "crossfade",
-    }
-    extra = {k: v for k, v in metadata.items() if k not in known_keys}
-
-    exaggeration = _parse_optional_float(metadata, "exaggeration")
-    cfg_weight = _parse_optional_float(metadata, "cfg_weight")
-    flow_steps = _parse_optional_int(metadata, "flow_steps")
-    cfg_alpha = _parse_optional_float(metadata, "cfg_alpha")
-    temperature = _parse_optional_float(metadata, "temperature")
-    top_p = _parse_optional_float(metadata, "top_p")
-    min_p = _parse_optional_float(metadata, "min_p")
-    repetition_penalty = _parse_optional_float(metadata, "repetition_penalty")
-    segment_gap = _parse_optional_int(metadata, "segment_gap")
-    crossfade = _parse_optional_int(metadata, "crossfade")
-
-    # Defaults that segments inherit from frontmatter
-    seg_defaults: dict = {}
-    if metadata.get("instruct"):
-        seg_defaults["instruct"] = metadata["instruct"]
-    for part in _INSTRUCT_PARTS:
-        if metadata.get(part):
-            seg_defaults[part] = metadata[part]
-    if metadata.get("language"):
-        seg_defaults["language"] = metadata["language"]
-    if metadata.get("voice"):
-        seg_defaults["voice"] = metadata["voice"]
-    if exaggeration is not None:
-        seg_defaults["exaggeration"] = exaggeration
-    if cfg_weight is not None:
-        seg_defaults["cfg_weight"] = cfg_weight
-    if flow_steps is not None:
-        seg_defaults["flow_steps"] = flow_steps
-    if cfg_alpha is not None:
-        seg_defaults["cfg_alpha"] = cfg_alpha
-    if temperature is not None:
-        seg_defaults["temperature"] = temperature
-    if top_p is not None:
-        seg_defaults["top_p"] = top_p
-    if min_p is not None:
-        seg_defaults["min_p"] = min_p
-    if repetition_penalty is not None:
-        seg_defaults["repetition_penalty"] = repetition_penalty
-    if segment_gap is not None:
-        seg_defaults["segment_gap"] = segment_gap
-    if crossfade is not None:
-        seg_defaults["crossfade"] = crossfade
-
-    segments = _parse_segments(body, seg_defaults)
-    text = " ".join(seg.text for seg in segments) if segments else strip_markdown(body)
-
-    # Doc-level instruct: explicit wins, otherwise compose from parts
-    doc_instruct = metadata.get("instruct")
-    if doc_instruct is None:
-        doc_instruct = compose_instruct(
-            metadata.get("accent"),
-            metadata.get("personality"),
-            metadata.get("speed"),
-            metadata.get("emotion"),
-        )
-
-    return TTSDocument(
-        text=text,
-        language=metadata.get("language"),
-        voice=metadata.get("voice"),
-        engine=metadata.get("engine"),
-        instruct=doc_instruct,
-        accent=metadata.get("accent"),
-        personality=metadata.get("personality"),
-        speed=metadata.get("speed"),
-        emotion=metadata.get("emotion"),
-        exaggeration=exaggeration,
-        cfg_weight=cfg_weight,
-        flow_steps=flow_steps,
-        cfg_alpha=cfg_alpha,
-        temperature=temperature,
-        top_p=top_p,
-        min_p=min_p,
-        repetition_penalty=repetition_penalty,
-        segment_gap=segment_gap,
-        crossfade=crossfade,
-        extra=extra,
-        segments=segments,
-    )
+    return _fn(body, defaults)
