+++ tools/file_exemptions.txt
@@ -4,9 +4,4 @@
 src/voicecli/cli.py https://github.com/Roxabi/voiceCLI/issues/80
 src/voicecli/api.py https://github.com/Roxabi/voiceCLI/issues/81
 src/voicecli/stt_daemon.py https://github.com/Roxabi/voiceCLI/issues/82
-src/voicecli/overlay.py https://github.com/Roxabi/voiceCLI/issues/83
-src/voicecli/markdown.py https://github.com/Roxabi/voiceCLI/issues/83
-src/voicecli/translate.py https://github.com/Roxabi/voiceCLI/issues/83
-src/voicecli/daemon.py https://github.com/Roxabi/voiceCLI/issues/83
-src/voicecli/nats/tts_adapter.py https://github.com/Roxabi/voiceCLI/issues/83
-src/voicecli/nats/stt_adapter.py https://github.com/Roxabi/voiceCLI/issues/83
+src/voicecli/daemon_protocol.py https://github.com/Roxabi/voiceCLI/issues/83
